+++ CHANGELOG.md
@@ -3,6 +3,7 @@ English | [简体中文](./CHANGELOG_CN.md)
 ## 3.10.1 (2021-12-xx)
 
 - `Feat(Network)` Pretty output format for JSON response. (issue #486)
+- `Fix(Style)` Avoid panel scaling with the web page When `initial-scale !== 1`.
 
 
 ## 3.10.0 (2021-12-17)